backend/pyproject.toml
[tool.poetry]
name = "copygram"
version = "0.1.0"
description = ""
authors = ["kvydyk <kvydyk@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
absl-py = "2.1.0"
amqp = "5.2.0"
asgiref = "3.8.1"
asttokens = "2.4.1"
astunparse = "1.6.3"
async-timeout = "4.0.3"
attrs = "23.2.0"
autobahn = "23.6.2"
Automat = "22.10.0"
beautifulsoup4 = "4.12.2"
billiard = "4.2.0"
blosc2 = "2.7.1"
celery = "5.4.0"
celery-singleton = "0.3.1"
certifi = "2023.7.22"
cffi = "1.16.0"
channels = "4.0.0"
channels-redis = "4.2.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.3.0"
colorama = "0.4.6"
comm = "0.2.2"
constantly = "23.10.4"
cron-descriptor = "1.4.3"
cryptography = "41.0.5"
debugpy = "1.8.8"
decorator = "5.1.1"
defusedxml = "0.8.0rc2"
Django = "5.1.3"
django-bootstrap5 = "23.4"
django-celery-beat = "2.7.0"
django-celery-results = "2.5.1"
django-debug-toolbar = "4.2.0"
django-elasticsearch-dsl = "8.0"
django-extensions = "3.2.3"
django-filter = "23.4"
django-formtools = "2.4.1"
django-oauth-toolkit = "3.0.1"
django-otp = "1.3.0"
django-phonenumber-field = "7.2.0"
django-phonenumbers = "1.0.1"
django-rename-app = "0.1.7"
django-taggit = "5.0.1"
django-timezone-field = "6.1.0"
django-tree-queries = "0.19.0"
django-two-factor-auth = "1.15.5"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.0"
drf-social-oauth2 = "3.1.0"
drf-spectacular = "0.27.2"
drf-yasg = "1.21.7"
elastic-transport = "8.15.1"
elasticsearch = "8.16.0"
elasticsearch-dsl = "8.16.0"
executing = "2.0.1"
Faker = "23.2.1"
flake8 = "7.1.1"
flatbuffers = "24.3.25"
flower = "2.0.1"
gast = "0.6.0"
google-pasta = "0.2.0"
grpcio = "1.67.1"
gunicorn = "22.0.0"
h11 = "0.14.0"
h5py = "3.12.1"
humanize = "4.9.0"
hyperlink = "21.0.0"
idna = "3.4"
incremental = "22.10.0"
inflection = "0.5.1"
ipykernel = "6.29.5"
ipython = "8.25.0"
jedi = "0.19.1"
joblib = "1.4.2"
jsonschema = "4.22.0"
jsonschema-specifications = "2023.12.1"
jupyter_client = "8.6.3"
jupyter_core = "5.7.2"
jwcrypto = "1.5.6"
keras = "3.6.0"
kombu = "5.3.4"
libclang = "18.1.1"
Markdown = "3.7"
markdown-it-py = "3.0.0"
MarkupSafe = "2.1.5"
matplotlib-inline = "0.1.7"
mccabe = "0.7.0"
mdurl = "0.1.2"
ml-dtypes = "0.4.1"
msgpack = "1.0.7"
namex = "0.0.8"
ndindex = "1.9.2"
nest-asyncio = "1.6.0"
numexpr = "2.10.1"
numpy = "2.0.2"
oauthlib = "3.2.2"
opt_einsum = "3.4.0"
optree = "0.13.0"
packaging = "23.2"
pandas = "2.2.3"
parso = "0.8.4"
pexpect = "4.9.0"
phonenumbers = "8.13.26"
Pillow = "10.1.0"
platformdirs = "4.3.6"
plotly = "5.18.0"
prometheus-client = "0.19.0"
prompt-toolkit = "3.0.41"
protobuf = "5.28.3"
psutil = "6.1.0"
psycopg2-binary = "2.9.9"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
py-cpuinfo = "9.0.0"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pycodestyle = "2.12.1"
pycparser = "2.21"
pyflakes = "3.2.0"
Pygments = "2.17.2"
PyJWT = "2.8.0"
pyOpenSSL = "23.3.0"
pypng = "0.20220715.0"
python-crontab = "3.0.0"
python-dateutil = "2.8.2"
python-dotenv = "1.0.1"
python3-openid = "3.2.0"
pytz = "2023.3.post1"
PyYAML = "6.0.1"
pyzmq = "26.2.0"
qrcode = "7.4.2"
redis = "5.2.0"
referencing = "0.35.1"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
rich = "13.9.4"
rpds-py = "0.18.1"
scikit-learn = "1.5.2"
scipy = "1.14.1"
service-identity = "24.1.0"
setuptools = "70.0.0"
six = "1.16.0"
social-auth-app-django = "5.4.0"
social-auth-core = "4.5.0"
soupsieve = "2.5"
sqlparse = "0.4.4"
stack-data = "0.6.3"
tables = "3.10.1"
tenacity = "8.2.3"
tensorboard = "2.18.0"
tensorboard-data-server = "0.7.2"
tensorflow = "2.18.0"
termcolor = "2.5.0"
threadpoolctl = "3.5.0"
tornado = "6.3.3"
traitlets = "5.14.3"
Twisted = "23.10.0"
txaio = "23.1.1"
typing_extensions = "4.8.0"
tzdata = "2023.3"
uritemplate = "4.1.1"
urllib3 = "2.1.0"
uvicorn = "0.30.1"
vine = "5.1.0"
wcwidth = "0.2.10"
websockets = "12.0"
Werkzeug = "3.1.3"
wheel = "0.45.0"
wrapt = "1.16.0"
pydot = "^3.0.2"
coverage = "^7.6.8"
pytest-django = "^4.9.0"
pytest-asyncio = "^0.24.0"
boto3 = "^1.35.74"
django-storages = "^1.14.4"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "copygram.settings"
# -- recommended but optional:
python_files = ["tests.py"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
